Youtube section button navigate and design almost done
