+++ src/Components/InstagramSlider.jsx
@@ -1,0 +1,11 @@
+import React from 'react'
+
+const InstagramSlider = () => {
+  return (
+    <>
+        
+    </>
+  )
+}
+
+export default InstagramSlider
